Add Auth0-specific scope claim support
Auth0 deviates from the normal scope claim abbreviation of "scp" to the more expanded "scope". To support this the IsInScope() call of the functiontokenvalueprovider must be overwritten.

Since Auth0 implements the OpenIdConnectConfiguration protocol the instantiation of the Auth0Options object autonegotiates this from the published endpoint reducing client overhead.

Fix spelling error of TokenSigningKeyOptions class