Add support for list of scopes with OR operator for Auth0
In order to support migration of scope names in an API without hard, breaking changes to required scopes we need to support multiple scopes being allowed in an "any-is-authorized" fashion.

This may be extended to support arbitrary boolean logic in the future.

For SigningKeyValueProviders other than Auth0 only 1 scope is supported.